requirements.txt
# Core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
pydantic>=2.0
pyyaml>=6.0

# LLM Providers
openai>=1.50.0
anthropic>=0.39.0
ollama>=0.4.0

# Memory / Vector Store
chromadb>=0.5.0

# Scheduling
apscheduler>=3.10.0

# Web / Dashboard
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9

# Browser Automation
playwright>=1.48.0

# Email
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Notifications
python-telegram-bot>=21.0

# Web Scraping / Search
beautifulsoup4>=4.12.0
httpx>=0.27.0
duckduckgo-search>=6.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
